refactor: separate index lists into independent headers
