+++ test/algorithms/synth_engines/xag_synth.cpp
@@ -2,11 +2,12 @@
 
 #include <kitty/kitty.hpp>
 
-#include <mockturtle/algorithms/synth_engines/xag_synth.hpp>
 #include <mockturtle/algorithms/simulation.hpp>
+#include <mockturtle/algorithms/synth_engines/xag_synth.hpp>
 #include <mockturtle/networks/aig.hpp>
 #include <mockturtle/networks/xag.hpp>
 #include <mockturtle/utils/index_list/index_list.hpp>
+#include <mockturtle/utils/index_list/list_simulator.hpp>
 
 using namespace mockturtle;
 
@@ -50,7 +51,7 @@ TEST_CASE( "XAIG synththesizer - projections", "[xag_synth]" )
   {
     kitty::create_nth_var( proj_fns[i], i );
   }
-  
+
   kitty::create_from_hex_string( onset, "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAFF" );
   kitty::create_from_hex_string( careset, "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFF00" );
   kitty::ternary_truth_table<TT> func0( onset, careset );
@@ -66,3 +67,45 @@ TEST_CASE( "XAIG synththesizer - projections", "[xag_synth]" )
   raw = { 7, 1, 0, 3 };
   CHECK( raw == list.raw() );
 }
+
+template<uint32_t NumVars>
+void test_xag_n_input_functions()
+{
+  xag_synth_stats st;
+  xag_synth_decompose<NumVars> engine( st );
+  using TT = kitty::static_truth_table<NumVars>;
+  TT onset, careset;
+  large_xag_index_list list;
+
+  list_simulator<xag_index_list<true>, TT> sim;
+  std::vector<TT> divisor_functions;
+  TT tmp;
+  for ( auto i = 0; i < NumVars; ++i )
+  {
+    kitty::create_nth_var( tmp, i );
+    divisor_functions.emplace_back( tmp );
+  }
+  std::vector<TT const*> xs_r;
+  for ( auto i = 0u; i < NumVars; ++i )
+  {
+    xs_r.emplace_back( &divisor_functions[i] );
+  }
+  do
+  {
+    kitty::ternary_truth_table<TT> tt( onset );
+    engine( tt );
+    auto index_list = engine.get_list();
+
+    sim( index_list, xs_r );
+    TT res;
+    sim.get_simulation_inline( res, index_list, xs_r, index_list.po_at( 0 ) );
+    CHECK( onset == res );
+
+    kitty::next_inplace( onset );
+  } while ( !kitty::is_const0( onset ) );
+}
+
+// TEST_CASE( "XAIG synththesizer - 3 input functions", "[xag_synth]" )
+//{
+//   test_xag_n_input_functions<3>();
+// }
